Create AndroidInput interface
