+++ backends/gdx-backend-android/src/com/badlogic/gdx/backends/android/AndroidApplication.java
@@ -118,7 +118,7 @@ public class AndroidApplication extends Activity implements AndroidApplicationBa
 		setApplicationLogger(new AndroidApplicationLogger());
 		graphics = new AndroidGraphics(this, config, config.resolutionStrategy == null ? new FillResolutionStrategy()
 			: config.resolutionStrategy);
-		input = AndroidInputFactory.newAndroidInput(this, this, graphics.view, config);
+		input = createInput(this, this, graphics.view, config);
 		audio = createAudio(this, config);
 		this.getFilesDir(); // workaround for Android bug #10515463
 		files = new AndroidFiles(this.getAssets(), this.getFilesDir().getAbsolutePath());
@@ -176,7 +176,7 @@ public class AndroidApplication extends Activity implements AndroidApplicationBa
 		
 		// detect an already connected bluetooth keyboardAvailable
 		if (getResources().getConfiguration().keyboard != Configuration.KEYBOARD_NOKEYS)
-			this.getInput().keyboardAvailable = true;
+			input.setKeyboardAvailable(true);
 	}
 
 	protected FrameLayout.LayoutParams createLayoutParams () {
@@ -301,6 +301,11 @@ public class AndroidApplication extends Activity implements AndroidApplicationBa
 	}
 
 	@Override
+	public AndroidInput getInput () {
+		return input;
+	}
+
+	@Override
 	public Files getFiles () {
 		return files;
 	}
@@ -308,11 +313,6 @@ public class AndroidApplication extends Activity implements AndroidApplicationBa
 	@Override
 	public Graphics getGraphics () {
 		return graphics;
-	}
-
-	@Override
-	public AndroidInput getInput () {
-		return input;
 	}
 
 	@Override
@@ -364,7 +364,7 @@ public class AndroidApplication extends Activity implements AndroidApplicationBa
 		super.onConfigurationChanged(config);
 		boolean keyboardAvailable = false;
 		if (config.hardKeyboardHidden == Configuration.HARDKEYBOARDHIDDEN_NO) keyboardAvailable = true;
-		input.keyboardAvailable = keyboardAvailable;
+		input.setKeyboardAvailable(keyboardAvailable);
 	}
 
 	@Override
@@ -499,6 +499,11 @@ public class AndroidApplication extends Activity implements AndroidApplicationBa
 
 	@Override
 	public AndroidAudio createAudio (Context context, AndroidApplicationConfiguration config) {
-		return new DefaultAndroidAudio(context, config);
+		return new AndroidAudioImpl(context, config);
+	}
+
+	@Override
+	public AndroidInput createInput (Application activity, Context context, Object view, AndroidApplicationConfiguration config) {
+		return AndroidInputFactory.newAndroidInput(this, this, graphics.view, config);
 	}
 }
