+++ backends/gdx-backend-android/src/com/badlogic/gdx/backends/android/AndroidMouseHandler.java
@@ -16,14 +16,11 @@
 
 package com.badlogic.gdx.backends.android;
 
-import android.content.Context;
 import android.view.InputDevice;
 import android.view.MotionEvent;
-import android.view.View;
 
 import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.Input.Buttons;
-import com.badlogic.gdx.backends.android.AndroidInput.TouchEvent;
+import com.badlogic.gdx.backends.android.AndroidInputImpl.TouchEvent;
 
 /** Mouse handler for devices running Android >= 3.1.
  * 
@@ -32,7 +29,7 @@ public class AndroidMouseHandler {
 	private int deltaX = 0;
 	private int deltaY = 0;
 
-	public boolean onGenericMotion (MotionEvent event, AndroidInput input) {
+	public boolean onGenericMotion (MotionEvent event, AndroidInputImpl input) {
 		if ((event.getSource() & InputDevice.SOURCE_CLASS_POINTER) == 0) return false;
 
 		final int action = event.getAction() & MotionEvent.ACTION_MASK;
@@ -78,7 +75,7 @@ public class AndroidMouseHandler {
 		Gdx.app.log("AndroidMouseHandler", "action " + actionStr);
 	}
 
-	private void postTouchEvent (AndroidInput input, int type, int x, int y, int scrollAmount, long timeStamp) {
+	private void postTouchEvent (AndroidInputImpl input, int type, int x, int y, int scrollAmount, long timeStamp) {
 		TouchEvent event = input.usedTouchEvents.obtain();
 		event.timeStamp = timeStamp;
 		event.x = x;
